langchain-tutorials/3-TUI_LangGraph_agent_PostgreSQL_memory/python/requirements.txt
python-dotenv==1.0.1
langchain==0.2.14
langchain-core==0.2.33
langchain-community==0.2.12
langchain-openai==0.1.22
langgraph==0.2.4
langgraph-checkpoint-postgres==1.0.3
psycopg==3.2.1
psycopg-pool==3.2.2
rich==13.7.1